refactor: use the transfer algorithms and source/sink types from boost.http
push_to and the source/sink type-erased wrappers (any_buffer_source,
any_buffer_sink, any_read_source) plus the BufferSource concept moved from
capy to boost.http. Point burl's includes and qualifications at their new
home: <boost/http/io/...>, <boost/http/concept/buffer_source.hpp>, and
http:: instead of capy:: for those symbols.
