+++ include/boost/burl/request_body.hpp
@@ -12,7 +12,7 @@
 
 #include <boost/burl/detail/config.hpp>
 
-#include <boost/capy/io/any_buffer_sink.hpp>
+#include <boost/http/io/any_buffer_sink.hpp>
 #include <boost/capy/io_task.hpp>
 
 #include <concepts>
@@ -44,7 +44,7 @@ namespace burl
         @ref body_from_tag.
 */
 template<typename T>
-concept RequestBody = requires(T const& t, capy::any_buffer_sink& sink) {
+concept RequestBody = requires(T const& t, http::any_buffer_sink& sink) {
     { t.content_type() } -> std::same_as<std::optional<std::string>>;
     { t.content_length() } -> std::same_as<std::optional<std::uint64_t>>;
     { t.write(sink) } -> std::same_as<capy::io_task<>>;
