+++ src/string.cpp
@@ -48,7 +48,7 @@ public:
     }
 
     capy::io_task<>
-    write(capy::any_buffer_sink& sink) const
+    write(http::any_buffer_sink& sink) const
     {
         auto [ec, n] =
             co_await sink.write_eof(
@@ -80,7 +80,7 @@ public:
     }
 
     capy::io_task<>
-    write(capy::any_buffer_sink& sink) const
+    write(http::any_buffer_sink& sink) const
     {
         auto [ec, n] = co_await sink.write_eof(
             capy::make_buffer(body_));
