+++ test/unit/detail/send_file.cpp
@@ -13,7 +13,7 @@
 #include <boost/burl/error.hpp>
 
 #include <boost/capy/ex/run_async.hpp>
-#include <boost/capy/test/buffer_sink.hpp>
+#include <boost/http/test/buffer_sink.hpp>
 #include <boost/corosio/io_context.hpp>
 
 #include "../temp_file.hpp"
@@ -32,11 +32,11 @@ class send_file_test
     static std::error_code
     run(std::filesystem::path const& path,
         std::uint64_t size,
-        capy::test::buffer_sink& bs)
+        http::test::buffer_sink& bs)
     {
         corosio::io_context ioc;
         std::error_code ret;
-        capy::any_buffer_sink sink(&bs);
+        http::any_buffer_sink sink(&bs);
         capy::run_async(
             ioc.get_executor(),
             [&](capy::io_result<> r) { ret = r.ec; })
@@ -52,7 +52,7 @@ public:
         std::string const contents = "hello file body";
         temp_file tmp(contents);
 
-        capy::test::buffer_sink bs;
+        http::test::buffer_sink bs;
         auto ec = run(tmp.path, contents.size(), bs);
 
         BOOST_TEST(!ec);
@@ -65,7 +65,7 @@ public:
         std::string const contents = "hello";
         temp_file tmp(contents);
 
-        capy::test::buffer_sink bs;
+        http::test::buffer_sink bs;
         auto ec = run(tmp.path, contents.size() + 10, bs);
 
         BOOST_TEST(ec == error::file_changed);
@@ -77,7 +77,7 @@ public:
         std::string const contents = "hello file body";
         temp_file tmp(contents);
 
-        capy::test::buffer_sink bs;
+        http::test::buffer_sink bs;
         auto ec = run(tmp.path, 5, bs);
 
         BOOST_TEST(!ec);
